Add files via upload
This is my second attempt at going through the full instructions. All of the code looks identical to what is in the book, but there are several (seven) problems that I have been trying to debug. I am leaving this code here, and then I am deleting it and starting over for my third attempt. I will load that up by the end of the night